Add files via upload
The Final Project.docx is a description of the agile principals I have learned in the course.  The Scrum-agile approach PowerPoint Presentation is a presentation to potential companies using the the waterfall design.  This presentation is meant to identify Scrum-agile practices and contrast between agile and waterfall.  The presentation is intended to show companies the power of agile development.